attached_assets/requirements.txt
streamlit
pytesseract
PyMuPDF
Pillow
